src/requirements.txt
Flask==2.3.1
python-dotenv==1.0.0
stem==1.8.0
